+++ storedimo/database/migrations/2024_04_16_171934_create_productos.php
@@ -13,24 +13,27 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('productos', function (Blueprint $table) {
-            $table->increments('id_producto');
-            $table->string('nombre_producto')->nullable();
-            $table->unsignedInteger('id_categoria')->nullable();
-            $table->integer('precio_unitario')->nullable();
-            $table->integer('precio_detal')->nullable();
-            $table->integer('precio_por_mayor')->nullable();
-            $table->string('descripcion')->nullable();
-            $table->integer('stock_minimo')->nullable();
-            $table->unsignedInteger('id_estado')->nullable();
-            $table->string('tamano')->nullable();
-            $table->integer('cantidad')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
+        if (!Schema::hasTable('productos')){
+            Schema::create('productos', function (Blueprint $table) {
+                $table->increments('id_producto');
+                $table->string('nombre_producto')->nullable();
+                $table->unsignedInteger('id_categoria')->nullable();
+                $table->integer('precio_unitario')->nullable();
+                $table->integer('precio_detal')->nullable();
+                $table->integer('precio_por_mayor')->nullable();
+                $table->string('descripcion')->nullable();
+                $table->integer('stock_minimo')->nullable();
+                $table->unsignedInteger('id_estado')->nullable();
+                $table->string('tamano')->nullable();
+                $table->integer('cantidad')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('id_categoria')->references('id_categoria')->on('categorias');
-            $table->foreign('id_estado')->references('id_estado')->on('estados');
-        });
+                $table->foreign('id_categoria')->references('id_categoria')->on('categorias');
+                $table->foreign('id_estado')->references('id_estado')->on('estados');
+            });
+        }
+        
     }
 
     /**
@@ -40,6 +43,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('productos');
+        if (Schema::hasTable('productos')){
+            Schema::dropIfExists('productos');
+        }
     }
 };
