+++ storedimo/database/migrations/2025_01_09_172043_create_usuarios.php
@@ -13,21 +13,24 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('usuarios', function (Blueprint $table) {
-            $table->increments('id_usuario');
-            $table->string('nombre_usuario')->nullable();
-            $table->string('apellido_usuario')->nullable();
-            $table->string('usuario')->nullable();
-            $table->string('identificacion')->nullable();
-            $table->string('clave')->nullable();
-            $table->unsignedInteger('id_estado')->nullable();
-            $table->unsignedInteger('id_rol')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
+        if (!Schema::hasTable('usuarios')) {
+            Schema::create('usuarios', function (Blueprint $table) {
+                $table->increments('id_usuario');
+                $table->string('nombre_usuario')->nullable();
+                $table->string('apellido_usuario')->nullable();
+                $table->string('usuario')->nullable();
+                $table->string('identificacion')->nullable();
+                $table->string('clave')->nullable();
+                $table->unsignedInteger('id_estado')->nullable();
+                $table->unsignedInteger('id_rol')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('id_estado')->references('id_estado')->on('estados');
-            $table->foreign('id_rol')->references('id')->on('roles');
-        });
+                $table->foreign('id_estado')->references('id_estado')->on('estados');
+                $table->foreign('id_rol')->references('id')->on('roles');
+            });
+        }
+        
     }
 
     /**
@@ -37,6 +40,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('usuarios');
+        if (Schema::hasTable('usuarios')) {
+            Schema::dropIfExists('usuarios');
+        }
     }
 };
