+++ storedimo/database/migrations/2025_02_12_141417_create_personas.php
@@ -13,28 +13,31 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('personas', function (Blueprint $table) {
-            $table->increments('id_persona');
-            $table->unsignedInteger('id_tipo_persona')->nullable();
-            $table->unsignedInteger('id_tipo_documento')->nullable();
-            $table->string('identificacion')->nullable();
-            $table->string('nombres_persona')->nullable();
-            $table->string('apellidos_persona')->nullable();
-            $table->string('numero_telefono')->nullable();
-            $table->string('celular')->nullable();
-            $table->string('email')->nullable();
-            $table->unsignedInteger('id_genero')->nullable();
-            $table->string('direccion')->nullable();
-            $table->unsignedInteger('id_estado')->nullable();
+        if (!Schema::hasTable('personas')) {
+            Schema::create('personas', function (Blueprint $table) {
+                $table->increments('id_persona');
+                $table->unsignedInteger('id_tipo_persona')->nullable();
+                $table->unsignedInteger('id_tipo_documento')->nullable();
+                $table->string('identificacion')->nullable();
+                $table->string('nombres_persona')->nullable();
+                $table->string('apellidos_persona')->nullable();
+                $table->string('numero_telefono')->nullable();
+                $table->string('celular')->nullable();
+                $table->string('email')->nullable();
+                $table->unsignedInteger('id_genero')->nullable();
+                $table->string('direccion')->nullable();
+                $table->unsignedInteger('id_estado')->nullable();
 
-            $table->timestamps();
-            $table->softDeletes();
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('id_tipo_persona')->references('id_tipo_persona')->on('tipo_persona');
-            $table->foreign('id_tipo_documento')->references('id_tipo_documento')->on('tipo_documento');
-            $table->foreign('id_genero')->references('id_genero')->on('generos');
-            $table->foreign('id_estado')->references('id_estado')->on('estados');
-        });
+                $table->foreign('id_tipo_persona')->references('id_tipo_persona')->on('tipo_persona');
+                $table->foreign('id_tipo_documento')->references('id_tipo_documento')->on('tipo_documento');
+                $table->foreign('id_genero')->references('id_genero')->on('generos');
+                $table->foreign('id_estado')->references('id_estado')->on('estados');
+            });
+        }
+        
     }
 
     /**
@@ -44,6 +47,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('personas');
+        if (Schema::hasTable('personas')) {
+            Schema::dropIfExists('personas');
+        }
     }
 };
